### 2019-01-31 #1223 @whedon
[PRE REVIEW]: Python Sorted Containers
**Submitting author:** @grantjenks (<a href="http://orcid.org/0000-0001-5010-405X">Grant Jenks</a>)
**Repository:** <a href="https://github.com/grantjenks/python-sortedcontainers" target ="_blank">https://github.com/grantjenks/python-sortedcontainers</a>
**Version:** v2.1.0
