### 2019-01-31 #1223 @danielskatz
This software looks useful, even though there's a problem compiling the paper.  You can make changes to the source, then enter ```@whedon compile pdf``` as a new comment here to try to fix it.

Another issue I have is that it's not clear to me that this is research software according to the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html):
> JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library, or similar.

Can you help me understand this?
