### 2019-02-02 #1223 @grantjenks
Regarding the guidelines, I think it fits under:

```
supports the functioning of research instruments or the execution of research experiments
```

But I may be wrong. I had JOSS suggested to me by the creator of IPython. And I received a GitHub Issue asking how to cite the project. Consider some of the users:

* Zipline, an algorithmic trading library from Quantopian
* Angr, a binary analysis platform from UC Santa Barbara
* Trio, an async I/O library
* Dask Distributed, a distributed computation library supported by Continuum Analytics
* Astropy, a community Python library for astronomy

Does that satisfy?

Also, I looked through GitHub issues trying to understand how the PDF is generated:

* https://github.com/openjournals/joss/issues/370
* https://github.com/openjournals/joss/issues/379
* https://github.com/openjournals/whedon
* https://gist.github.com/mhucka/5ecb18aa2fd088c055de07366f70135e
* https://github.com/openbases/builder-pdf

But I was unable to reproduce the problem. (Maybe those links will be useful for a future visitor). There seems to be a "paper.md" file that is missing. Is that something I generate? I'm happy to accommodate, just not sure what's necessary. I could see how the project README may not be as useful for a paper as say the [Performance Comparison](http://www.grantjenks.com/docs/sortedcontainers/performance.html), [Implementation Details](http://www.grantjenks.com/docs/sortedcontainers/implementation.html), or [Performance at Scale](http://www.grantjenks.com/docs/sortedcontainers/performance-scale.html) materials.
