### 2019-02-03 #1223 @arfon
Hi @grantjenks - please take a look at the [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html) and prepare a `paper.md` and `paper.bib` file and include them in your repository (https://github.com/grantjenks/python-sortedcontainers). Once you've uploaded them, Whedon (the robot here) will be able to compile the paper.

In your paper, please be sure to address each of the points listed in the section [what should my paper contain?](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).
