### 2019-02-04 #1223 @arfon
> I _think_ that's fine, but want to hear from @arfon before we go on.

👍 I think that's OK too. Please go ahead and prepare a paper to accompany this submission @grantjenks.
