### 2019-02-13 5007b63
Simplify and improve performance of __reduce__
