### 2019-02-13 b5395b5
Simplify and improve performance of __reduce__
