### 2019-02-20 42efac4
Fixes typos in sortedlist docstrings
