### 2019-02-24 977c247
Add initial draft skeleton of paper for JOSS
