+++ docs/paper.md
@@ -1,0 +1,268 @@
+---
+title: 'Python Sorted Containers'
+tags:
+  - Python
+  - sorted
+  - list
+  - dictionary
+  - set
+authors:
+  - name: Grant Jenks
+    orcid: 0000-0000-0000-0000
+    affiliation: 1
+affiliations:
+ - name: Lyman Spitzer, Jr. Fellow, Princeton University
+   index: 1
+date: 23 February 2019
+bibliography: paper.bib
+---
+
+# Summary
+
+You’re probably more familiar with sorted collections than you realize.
+
+In the standard library we have heapq, bisect and queue.PriorityQueue but they
+don’t quite fill the gap. Behind the scenes, priority queue uses a heap
+implementation. Another common mistake is to think that collections.OrderedDict
+is a dictionary that maintains sort order but that’s not the case.
+
+I don’t always import sorted types. But when I do, I expect them in the
+standard library.
+
+And here’s why. Java, C++ and .NET have them. Python has broken into the top
+five of the TIOBE index but feels a bit more like PHP or Javascript in this
+regard.
+
+We also depend on external solutions: Sqlite in-memory indexes,
+pandas.DataFrame indexes, and Redis sorted sets. If you’ve ever issued a “zadd”
+command to Redis then you used a sorted collection.
+
+So what should be the API of sorted collection types in Python?
+
+Well, a sorted list should be a MutableSequence. Pretty close to the “list”
+API. But there’s a sort order constraint that must be satisfied by “setitem”
+and “insert” methods. Also should support a “key” argument like the “sorted”
+builtin function. Given sorted order, “bisect_right” and “bisect_left” methods
+make sense. You could also imagine an “add” method and “discard” method for
+elements. Kind of like a multi-set in other languages. I’d also expect
+“getitem”, “contains”, “count”, etc. to be faster than linear time.
+
+A sorted dictionary should be a MutableMapping. Pretty close to the dictionary
+API. But iteration yields items in sorted order. Also should support efficient
+positional indexing, something like a SequenceView.
+
+Sorted set should be a MutableSet. Pretty close to the “set” API. Sorted set
+should also be a Sequence like the “tuple” API to support efficient positional
+indexing.
+
+The chorus and the refrain from core developers is: “Look to the PyPI.” Which
+is good advice.
+
+So let’s talk about your options with a bit of software archaeology.
+
+Blist is the genesis of our story but it wasn’t really designed for sorted
+collections. It’s written in C and the innovation here is the “blist” data
+type. That’s a B-tree based replacement for CPython’s built-in list. Sorted
+list, sorted dictionary, and sorted set were built on top of this “blist” data
+type and it became the incumbent to beat. Also noteworthy is that the API was
+rather well thought out.
+
+There were some quirks; for example: the “pop” method returns the first element
+rather than the last element in the sorted list.
+
+SortedCollection is not a package. You can’t install this with “pip”. It’s
+simply a Python recipe that Raymond Hettinger linked from the Python
+docs. Couple innovations here though: it’s simple, it’s written in pure-Python,
+and maintains a parallel list of keys. So we have efficient support for that
+key-function parameter.
+
+This is bintrees. Still alive and kicking today. A few innovations here: it’s
+written with Cython support to improve performance and has a few different tree
+“backends.” You can create a red-black or AVL-tree depending on your
+needs. There’s also some notion of accessing the nodes themselves and
+customizing the tree traversal to slice by value rather than by index.
+
+Banyan had a very short life but adds another couple innovations: it’s
+incredibly fast and achieves that through C++ template meta-programming. It
+also has a feature called tree-augmentation that will let you store metadata at
+tree nodes. You can use this for interval trees if you need those.
+
+Finally there’s skiplistcollections. Couple significant things here: it’s
[…]
