+++ docs/paper.bib
@@ -1,0 +1,7 @@
+@manual{Python,
+    title = {Python Programming Language},
+    author = {Python Core Developers},
+    organization = {Python Software Foundation},
+    year = {1995},
+    url = {https://www.python.org/},
+}
