### 2019-03-05 9337b41
Initial draft of paper for JOSS
