+++ docs/paper.md
@@ -8,261 +8,78 @@ tags:
   - set
 authors:
   - name: Grant Jenks
-    orcid: 0000-0000-0000-0000
-    affiliation: 1
-affiliations:
- - name: Lyman Spitzer, Jr. Fellow, Princeton University
-   index: 1
-date: 23 February 2019
+    orcid: 0000-0001-5010-405X
+date: 5 March 2019
 bibliography: paper.bib
 ---
 
 # Summary
 
-You’re probably more familiar with sorted collections than you realize.
+The standard library of popular languages like C++, Java, and C# provide sorted
+container data types based on binary tree data structures. While Python has
+risen in popularity, the Standard Library still lacks these common data types.
+Part of the challenge has been Python's rich object model which makes binary
+trees implemented in Python slow in terms of both memory and processor
+usage. To overcome the overhead of the interpreter, C-extensions are used by
+the Python core developers. In doing so, flexibility is tradeoff for
+performance. The goal of the Python core developers is to provide the right set
+of high-level APIs so that algorithms and data structures can be implemented
+efficiently. The Python Sorted Containers library uses Python's high-level APIs
+to efficiently implement sorted container data structures.
 
-In the standard library we have heapq, bisect and queue.PriorityQueue but they
-don’t quite fill the gap. Behind the scenes, priority queue uses a heap
-implementation. Another common mistake is to think that collections.OrderedDict
-is a dictionary that maintains sort order but that’s not the case.
+Python's collections data structures are based on three data types: sequences,
+mappings, and sets. These data types are implemented and most commonly used as
+list, dictionary, and set objects. The Python Sorted Containers library
+introduces new variants of these three data types that are each sorted: sorted
+list, sorted dictionary, and sorted set. In each case, the original semantics
+are extended to preserve sorted order of the contained elements with respect to
+mutating operations. When unable to preserve the sorted order constraint, the
+functionality is either non-existent or an error is raised from the
+library. Python's "sorted" built-in function also supports a "key" parameter
+which specifies a callable used to extract a comparison key from elements. When
+initializing a sorted container data type, the key paramemter is likewise
+supported.
 
-I don’t always import sorted types. But when I do, I expect them in the
-standard library.
+Internally, Python Sorted Containers uses a list of sublists data structure
+that is like a B-tree contstrained to two levels of nodes. The maximum of each
+sublist is maintained in a separate list. To lookup an element, the list of
+maximums is bisected using the "bisect" module in the Standard Library. Using
+the bisected maximums index, the corresponding sublist is bisected to find the
+index of the desired element. To index the k'th element, a separate positional
+index, known as the "Jenks" index, is built and maintained. The positional
+index is like the order statistic of a binary tree densely packed into a
+list. By maintaining the size of the sublists as proportional to the
+$\sqrt[3]{n}$ the amortized time complexity of all operations is
+$O(\sqrt[3]{n})$. This bound works well for up to billions of elements which
+often exhausts all available memory.
 
-And here’s why. Java, C++ and .NET have them. Python has broken into the top
-five of the TIOBE index but feels a bit more like PHP or Javascript in this
-regard.
-
-We also depend on external solutions: Sqlite in-memory indexes,
-pandas.DataFrame indexes, and Redis sorted sets. If you’ve ever issued a “zadd”
-command to Redis then you used a sorted collection.
-
-So what should be the API of sorted collection types in Python?
-
-Well, a sorted list should be a MutableSequence. Pretty close to the “list”
-API. But there’s a sort order constraint that must be satisfied by “setitem”
-and “insert” methods. Also should support a “key” argument like the “sorted”
[…]
