+++ docs/paper.bib
@@ -1,7 +1,107 @@
-@manual{Python,
-    title = {Python Programming Language},
-    author = {Python Core Developers},
+@manual{ python,
+    title        = {Python Programming Language},
+    author       = {Python Core Developers},
     organization = {Python Software Foundation},
-    year = {1995},
-    url = {https://www.python.org/},
+    year         = {1995},
+    url          = {https://www.python.org/},
+}
+
+@inproceedings{ angr-shoshitaishvili2016state,
+    title     = {{SoK: (State of) The Art of War: Offensive Techniques in Binary Analysis}},
+    author    = {Shoshitaishvili, Yan and Wang, Ruoyu and Salls, Christopher and Stephens, Nick and Polino, Mario and Dutcher, Audrey and Grosen, John and Feng, Siji and Hauser, Christophe and Kruegel, Christopher and Vigna, Giovanni},
+    booktitle = {IEEE Symposium on Security and Privacy},
+    year      = {2016}
+}
+
+@article{ astropy:2013,
+    adsnote       = {Provided by the SAO/NASA Astrophysics Data System},
+    adsurl        = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A},
+    archiveprefix = {arXiv},
+    author        = {{Astropy Collaboration} and {Robitaille}, T.~P. and {Tollerud}, E.~J. and {Greenfield}, P. and {Droettboom}, M. and {Bray}, E. and {Aldcroft}, T. and {Davis}, M. and {Ginsburg}, A. and {Price-Whelan}, A.~M. and {Kerzendorf}, W.~E. and {Conley}, A. and {Crighton}, N. and {Barbary}, K. and {Muna}, D. and {Ferguson}, H. and {Grollier}, F. and {Parikh}, M.~M. and {Nair}, P.~H. and {Unther}, H.~M. and {Deil}, C. and {Woillez}, J. and {Conseil}, S. and {Kramer}, R. and {Turner}, J.~E.~H. and {Singer}, L. and {Fox}, R. and {Weaver}, B.~A. and {Zabalza}, V. and {Edwards}, Z.~I. and {Azalee Bostroem}, K. and {Burke}, D.~J. and {Casey}, A.~R. and {Crawford}, S.~M. and {Dencheva}, N. and {Ely}, J. and {Jenness}, T. and {Labrie}, K. and {Lim}, P.~L. and {Pierfederici}, F. and {Pontzen}, A. and {Ptak}, A. and {Refsdal}, B. and {Servillat}, M. and {Streicher}, O.},
+    doi           = {10.1051/0004-6361/201322068},
+    eid           = {A33},
+    eprint        = {1307.6212},
+    journal       = {\aap},
+    keywords      = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
+    month         = Oct,
+    pages         = {A33},
+    primaryclass  = {astro-ph.IM},
+    title         = {{Astropy: A community Python package for astronomy}},
+    volume        = 558,
+    year          = 2013,
+    bdsk-url-1    = {https://dx.doi.org/10.1051/0004-6361/201322068}
+}
+
+@article{ astropy:2018,
+    adsnote      = {Provided by the SAO/NASA Astrophysics Data System},
+    adsurl       = {https://ui.adsabs.harvard.edu/#abs/2018AJ....156..123T},
[…]
