+++ docs/paper.bib
@@ -105,3 +105,15 @@
     howpublished = "\url{http://www.zipline.io/}",
     note         = "[Online; accessed 5-March-2019]"
 }
+
+@article{ zenodo,
+    abstractnote  = {Python Sorted Container Types: SortedList, SortedDict, and SortedSet},
+    author        = {Grant Jenks},
+    date-modified = {2019-03-05 21:21:18 +0000},
+    doi           = {10.5281/zenodo.2584337},
+    month         = {Jul},
+    publisher     = {Zenodo},
+    title         = {grantjenks/python-sortedcontainers: v2.1.0},
+    year          = {2019},
+    bdsk-url-1    = {http://dx.doi.org/10.5281/zenodo.2584337}
+}
