### 2019-03-05 7d3160d
Remove unnecessary else
