### 2019-03-05 be06d8d
Pickle fix for SortedSet on py27
