### 2019-03-05 #1223 @grantjenks
I published the paper.md and paper.bib files this morning. I believe I covered each of the requirements as to "What should my paper contain?" It is ready to be un-paused. I believe that was pre-review. Let me see if I can get whedon to compile the PDF.
