+++ docs/paper.md
@@ -9,6 +9,7 @@ tags:
 authors:
   - name: Grant Jenks
     orcid: 0000-0001-5010-405X
+affiliations:
 date: 5 March 2019
 bibliography: paper.bib
 ---
