### 2019-03-07 #1223 @grantjenks
Hi all, I'm not sure what you want for "affiliations". I was not attending college when I developed the project. But I can put where I graduated. Should I put my company? Please advise.
