### 2019-03-07 #1223 @arfon
@grantjenks - you can have an affiliation of 'None', but the YAML header of the paper needs to match this example: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
