### 2019-03-07 96d5ce0
Add None affiliation
