### 2019-03-18 #1330 @whedon
[REVIEW]: Python Sorted Containers
**Submitting author:** @grantjenks (<a href="http://orcid.org/0000-0001-5010-405X">Grant Jenks</a>)
**Repository:** <a href="https://github.com/grantjenks/python-sortedcontainers" target ="_blank">https://github.com/grantjenks/python-sortedcontainers</a>
**Version:** v2.1.0
