### 2019-03-18 #1223 @jasonclark
@djmitche the review is set up at https://github.com/openjournals/joss-reviews/issues/1330. cc @grantjenks
