### 2019-03-23 #1330 @djmitche
@grantjenks:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I don't see these bits in the repository.  Have I missed them?

Typically (1) is handled with `CONTRIBUTORS.md` or a section in the README, and (2) and (3) are handled with GitHub issue templates.
