### 2019-03-23 #1330 @grantjenks
There's a page for that at http://www.grantjenks.com/docs/sortedcontainers/development.html It's the second-to-last link under the User Guide section of the README. I think it covers 1-3.
