### 2019-03-23 #1330 @djmitche
Indeed, I did miss that.  That definitely covers (1), but not so much (2) and (3).  It's probably a tiny thing, but just an indication that both support requests and bug reports are welcome on the github repo would be enough.
