### 2019-03-26 fe3d690
Add sections for Reporting Issues and Support to docs.
