+++ docs/index.rst
@@ -8,9 +8,9 @@
    performance-load
    performance-runtime
    performance-workload
-   implementation
    performance-scale
    development
+   implementation
    history
    sortedlist
    sorteddict
