+++ README.rst
@@ -221,7 +221,7 @@ Resources
 Sorted Containers License
 -------------------------
 
-Copyright 2014-2018 Grant Jenks
+Copyright 2014-2019 Grant Jenks
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
