### 2019-03-26 #1330 @grantjenks
@djmitche I have updated the docs at [Development and Support](http://www.grantjenks.com/docs/sortedcontainers/development.html) to cover (2) and (3).

There's two new sections at the top of the page: "Reporting Issues" and "Support". I've also tried to make it a bit more discoverable by updating the README's Documentation section with a Community Guide heading that links to "Development and Support". For now, I'd prefer not to use GitHub issue templates. I think the documentation provides sufficient guidelines for (2) and (3).

Thanks for the feedback and review!
