### 2019-04-02 29aeebf
Fix a warning under Python 3.7
`Warning: 'classifiers' should be a list, got type 'tuple'`

Related issue: https://bugs.python.org/issue19610
