### 2019-04-05 #1330 @grantjenks
I think that's done. I believe the DOI is: 10.5281/zenodo.2584337
Here's the Zenodo link: https://zenodo.org/record/2584337
And the DOI link: https://doi.org/10.5281/zenodo.2584337
