### 2019-04-18 b5743f8
suppress warning in test of deprecated function
