### 2019-04-18 6ee5d57
improve SortedList.__getitem__() performance for small slices
