### 2019-04-18 11a6663
improve SortedList.__getitem__() performance for small slices
