### 2019-05-18 #1330 @arfon
> @jasonclark hey, so this keeps showing up in my "assigned" queue despite there being no action required on my part. And @whedon seems to want it that way. Can we wrap this up?

Sorry @djmitche - Whedon periodically tries to fix assignments on review issues but it looks like it's been a little over-keen here.

@grantjenks - I've just emailed Jason a reminder. If we don't hear back from him in the next day or so then I'll finish up this submission for him.
