### 2019-05-21 #1330 @arfon
:wave: @grantjenks - a couple of final changes to the paper:

- Could you remove the last sentence about the Zenodo archive? We link to the archive directly from the paper already so this is superfluous.
- When you're mentioning projects (e.g. Astropy, Dask etc.) could you cite those packages in the paper? You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html). Also, note the Dask paper needs a DOI added to it.
