### 2019-06-03 626ed57
Add links to references in paper.
Updated Dask Distributed paper based on
http://conference.scipy.org/proceedings/scipy2015/matthew_rocklin.html
