+++ docs/paper.bib
@@ -77,7 +77,8 @@
     author    = {Rocklin, Matthew},
     title     = {Dask: Parallel Computation with Blocked algorithms and Task Scheduling},
     booktitle = {Proceedings of the 14th Python in Science Conference},
-    pages     = {130 - 136},
+    doi       = {10.25080/Majora-7b98e3ed-013},
+    pages     = {126 - 132},
     year      = {2015},
     editor    = {Kathryn Huff and James Bergstra}
 }
