### 2019-06-03 #1330 @grantjenks
@arfon Update:

1. I removed the last sentence and verified the article proof still includes the DOI.
2. I found the Dask Distributed DOI and included it.
3. I cited the projects in the paper based on your reference. I wasn't sure how the text would be rendered. Looks like it included just the year. I think that's OK but let me know if there's a better rendering.
4. Everyone in the bibliography now has a reference from the paper except astropy:2013. Should I remove that entry from paper.bib?

I'm excited to wrap this up today, if possible.
